requirements.txt
ariadne==0.14.1
marshmallow==3.14.1
requests==2.27.1
Flask==1.1.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.20
itsdangerous==2.0.1
attr==0.3.1
savalidation==0.4.1
Flask-Migrate==3.1.0
PyMySQL==1.0.2
PyJWT==2.3.0